+++ examples/CRISP/interfold.config.yaml
@@ -16,9 +16,6 @@ chains:
       bonding_registry:
         address: "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853"
         deploy_block: 14
-      bonding_registry:
-        address: "0xa513E6E4b8f2a923D98304ec87F64353C4D5C853"
-        deploy_block: 14
       slashing_manager:
         address: "0xCf7Ed3AccA5a467e9e704C703E8D87F634fB0Fc9"
         deploy_block: 12
